Create  reverse-of-given-input.java
Time Complexity:

The time complexity of this Java program is O(n), where n is the length of the input string. We iterate through the string characters once to reverse it.
Space Complexity:

The space complexity is O(n) because we create a character array of the same length as the input string to store the reversed string.