Create PalindromeChecker.java
In this program, you can input either a string or an integer, and it will determine whether the input is a palindrome. The Scanner class is used to read input from the user. Depending on the type of input (string or integer), it calls the corresponding isPalindrome function to check for a palindrome.

To test this program, compile and run it. Enter a string or an integer when prompted, and it will tell you if it's a palindrome.